Back to ints and fixed overflow

+++ ComputerVision/src/Robot.cpp
@@ -29,11 +29,11 @@ Robot::~Robot() {
 }
 
 // Setters and Getters
-void Robot::setAngle(float newAngle) {
+void Robot::setAngle(int newAngle) {
   Robot::angle = newAngle;
 }
 
-float Robot::getAngle() {
+int Robot::getAngle() {
   return Robot::angle;
 }
